modalgateway-Tops-1/modalgateway-Tops-1/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
python-dotenv==1.0.0
together>=1.2.0
openai==1.10.0
PyPDF2==3.0.1
python-docx==1.1.0
pydantic>=2.6.3
python-multipart==0.0.6
